ADM-809[frontend]rename file
