+++ frontend/e2e/pages/Home.ts
@@ -1,5 +1,5 @@
-import { importFlagAsBlockFile } from '../fixtures/add-flag-as-block-config-file';
-import { importProjectFromFile } from '../fixtures/hb-e2e-for-importing-file';
+import { importMultipleDoneProjectFromFile } from '../fixtures/importFile/multiple-done-config-file';
+import { importFlagAsBlockFile } from '../fixtures/importFile/add-flag-as-block-config-file';
 import { expect, Locator, Page } from '@playwright/test';
 export class HomePage {
   readonly page: Page;
@@ -24,11 +24,12 @@ export class HomePage {
   async createANewProject() {
     await this.createANewProjectButton.click();
   }
+
   async importMultipleDoneProjectFromFile() {
     await this.importProjectFromFileInput.setInputFiles({
       name: 'hb-e2e-test',
       mimeType: 'text/plain',
-      buffer: Buffer.from(JSON.stringify(importProjectFromFile)),
+      buffer: Buffer.from(JSON.stringify(importMultipleDoneProjectFromFile)),
     });
   }
